make notifications a component
