--- app/controllers/notifications.ts
@@ -1,17 +1,0 @@
-'use strict';
-export default function(app) {
-  app.controller('NotificationsCtrl', ['$scope', 'notificationService',
-    function($scope, notificationService) {
-      $scope.notifications = notificationService.notifications;
-      $scope.close = function(index) {
-        notificationService.notifications.splice(index, 1);
-      };
-      $scope.types = {
-        'error': 'danger',
-        'info': 'info',
-        'warning': 'warning',
-        'success': 'success'
-      };
-    }
-  ]);
-};
