+++ html/shared/navbar.html
@@ -30,7 +30,7 @@
     </div><!-- collapse -->
   </div><!-- container-fluid -->
   <div>
-    <ng-include src="'html/shared/notifications.html'"></ng-include>
+    <notifications></notifications>
   </div>
 </div><!-- navbar -->
 <div class="ph-navbar-shadow"><!-- #uglyhack: shadow z-index < subnav z-index -->
